requirements.txt
scapy>=2.5.0
python-nmap>=0.7.1
psutil>=5.9.0
rich>=13.7.0
tabulate>=0.9.0
pandas>=2.1.0
requests>=2.31.0
PyQt6>=6.6.1
pywin32>=306 